requirements.txt
absl-py==2.1.0
asttokens==2.0.5
backcall==0.2.0
Brotli==1.0.9
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.1
executing==0.8.3
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grad-cam==1.5.4
grpcio==1.67.1
huggingface-hub==0.29.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
loguru==0.7.2
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==6.1.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.15.1
pyparsing==3.1.4
PySnooper==1.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
stack-data==0.2.0
sympy==1.13.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.14
torch==2.1.2
torch-pruning==0.2.7
torchaudio==2.1.2
TorchSnooper==0.8
torchvision==0.16.2
tqdm==4.67.0
traitlets==5.14.3
ttach==0.0.3
typing_extensions==4.11.0
tzdata==2024.2
ultralytics==8.3.29
ultralytics-thop==2.0.11
urllib3==2.2.3
wcwidth==0.2.5
Werkzeug==3.0.6
wheel==0.44.0
win-inet-pton==1.1.0
win32-setctime==1.1.0
zipp==3.20.2
